% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvRLab5.R
\docType{package}
\name{AdvRLab5}
\alias{AdvRLab5}
\alias{AdvRLab5-package}
\title{AdvRLab5: A package we have to create in lab 3 of advanced R Programming}
\description{
AdvRLab5: A package we have to create in lab 3 of advanced R Programming
}
\section{TwitterWordCould function}{

This function creates a data frame which can be used to create a word cloud
}

