% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextMining_Twitter_API.R
\name{TwitterWordCould}
\alias{TwitterWordCould}
\title{TwiterWordCloud}
\arguments{
\item{word}{, is the tweet search word}

\item{min_freq}{, is the minimum frequency of the tweet search}

\item{max_number}{, is the maximum number of the tweet search}
}
\value{
df_Tesla , returns a data frame with the term and frequency
}
\description{
function create a data fram of the searched tweet including the frequency
}
